fix(ax3): extract the writer's task-ledger application into task_ops
writer/ops.rs was at 1162 production lines, over RULE-003's 1000-line
cap, and its apply_task_assignment was a 91-line RULE-002 hard
violation newly introduced by this branch.

Move the sole task-ledger application site (apply_task_message,
apply_task_assignment, apply_task_completion, apply_task_acknowledgement,
append_task_event, and their shared task_rejected/TASK_RECOVERY helpers)
into a new private writer::task_ops module with no behavior change, and
split apply_task_assignment/apply_task_completion/apply_task_acknowledgement
into smaller named helpers (load_existing_assignment_state,
refresh_task_assignment, insert_task_assignment,
acknowledge_completed_assignment, reject_if_another_task_is_active) so no
function in the split exceeds RULE-002's 80-line hard-violation
threshold. ops.rs keeps calling apply_task_message/apply_task_acknowledgement
through pub(super) re-exports from the new sibling module.
